Move MemoryTicketStore to API module
