+++ tmail-backend/jmap/extensions-api/src/main/scala/com/linagora/tmail/james/jmap/ticket/MemoryTicketStore.scala
@@ -1,0 +1,21 @@
+package com.linagora.tmail.james.jmap.ticket
+
+import reactor.core.scala.publisher.SMono
+
+import scala.collection.mutable
+
+class MemoryTicketStore extends TicketStore {
+  private val map: mutable.Map[TicketValue, Ticket] = mutable.Map()
+
+  override def persist(ticket: Ticket): SMono[Unit] =
+    SMono.fromCallable(() => map.put(ticket.value, ticket))
+      .`then`
+
+  override def retrieve(value: TicketValue): SMono[Ticket] = map.get(value)
+    .map(SMono.just)
+    .getOrElse(SMono.empty)
+
+  override def delete(ticketValue: TicketValue): SMono[Unit] =
+    SMono.fromCallable(() => map.remove(ticketValue))
+      .`then`
+}
